dashboard/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icarus-dashboard"
version = "0.1.0"
description = "Operations console for icarus-memory: wiki, activity, review queue"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "esaradev" }]

dependencies = [
    "icarus-memory>=0.3.0",
    "python-fasthtml>=0.6",
    "uvicorn>=0.27",
    "watchfiles>=0.21",
    "markdown-it-py>=3.0",
]

[project.scripts]
icarus-dashboard = "icarus_dashboard.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/icarus_dashboard"]
